refactor: 패키지명 단수로 통일 - constants -> constant
